Don't retry POST from OCM gateway to assisted installer
A previous patch changed the configuration of the OCM gateway so that
all requests to the assisted installer service are retried when the
connection is reset. If the retried request is a POST and the request
arrives to the backend before the connection reset it is possible to
have the corresponding object created twice. This is unlikely because
the connection resets come from the OpenShift router in front of the
backend, and that will not close a connection without sending at least a
503 error response. Anyhow we want to mitigate the possibility of
creating objects twice. To do so this patch changes the configuration of
the OCM gateway so that it doesn't retry POST requests.

Related: https://issues.redhat.com/browse/MGMT-11711
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>