assisted-installer: set release image in integration env
Override temporary the release images in integration, this is required
for the testing of assisted installer 4.9.0.

For integration, the default release image is embeded in the service in
integration environment, so once this testing is done, this PR needs to
be reverted.

If we want to keep a simillar setup, we'll need to update the default
images in the assisted service
https://github.com/openshift/assisted-service/tree/master/data.