Persist quay-mirror control file - stage
With the amount of deployments we do per day, we run in compare-tags
mode (aka deep sync) more than we would like (once a day). We have added
the possibility to specify the directory in which we can keep the
control file so that we can make sure we only run in compare-tags mode
once a day (it is VERY slow).

part of APPSRE-6102.

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>