Promote qontract-reconcile
The integration does not run in non-compare-tags mode (aka shallow
mode) in dry-run, as it made testing basic scenarios (e.g. compare tags
in dry-run mode) something that it could only be achieved by directly
modifying the code.

In a MR we don't want to compare tags, as that is something that is
completely outside app-interface, so we will only check for new tags.

xref: https://github.com/app-sre/qontract-reconcile/pull/2765

part of APPSRE-6102.

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>