quay-mirror no compare tags in stage
We want to mimick the old behaviour, dry-run didn't compare tags. It's
just not worth to compare tags in dry-run mode. Errors are going to
appear very far in the run, they won't be checked as they won't appear
soon and they will be fixed analyzed in prod where they either won't
affect the normal run or they will be analyzed as part of alerting.

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>