image-builder-crc: copy accounts_quotas to config_files
We want to add more config files, so let's rename the accounts_quotas config
map to a more generic name - config_files.

Actually, we cannot just rename the config map, since the deployment is still
using that, so this commit just copies it for the future use. We will remove
the old map when the switch is done.