image-builder-crc: copy accounts_quotas to config_files on prod
We want to add more config files, so let's rename the accounts_quotas config
map to a more generic name - config_files.

Actually, we cannot just rename the config map, since the deployment is still
using that, so this commit just copies it for the future use. We will remove
the old map when the switch is done.

This is the same commit as 1b69eaad2c9e0d00cb09d825d228f1a1967a9a5f but on
prod.