Switch stage fleet manager to AMS
This changes the source of truth of user account quota to AMS as opposed
to the current hard coded quota list. AMS supports a) reconciliation of
SKUs purchased by customers from Red Hat IT systems, and b) hard coded
granted via ocm-resources. The latter has been used to give quota to the
Red Hat umbrella organization.

Note that AMS stage and production environments are separate and do not
share the same quota.