Add mirrored OCP image of 4.11.5
To be able to do performance tests and not overload quay.io, we were
asked to work with a mirror for the images. This should add an image
mirrored by the pull-through mirror in
ocp-qe-01.savanna.lab.eng.rdu2.redhat.com:5015, and hopefully will have
a smaller footprint on quay.io usage.