feature: add "mbarcina" to the notifications and policies team
I am adding myself a few roles that are going to be required for me to
be able to work on these teams. I am not removing myself from Sources
yet because I will be performing a database upgrade for that team, and I
will still need access to the SASS files.