enable MR management for change-owners
when MR management is active, change-owners manages the self-serviceable and non-self-serviceable labels on MRs, if not, it even removes them. it is crucial that MR management is only activated when saas-file-owners is not running.

activating MR management also makes sure the informational comment about who can LGTM what is written to the MR and approval comments are processed. this results in bot/approved and bot/hold labels to be managed as well to give clear indication to gitlab-housekeeper, if an MR can be merged

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>