Update Obs API Route configuration for Telemeter
It matches the timeout and load balance strategy set in the RHOBS MST.

The idea is that the timeout should be controlled by Observatorium API and its downstream components. We expect some requests to take long to process (i.e. 2+ minutes).

The load balancer strategy is chosen so because Obs API is a proxy that sends requests downstream into a Kubernetes Service URL. Such service takes care of load balancing.