Bump swatch-tally-service replicas to 3 in stage
Prod intent
-----------
cpu request 2 -> 1
cpu limit 4.8 -> 1

This is based on the observation that the tally service doesn't
generally use more than 50% of given CPU.

memory request: 14000Mi -> 7000Mi
memory limit: 15488Mi -> 7000Mi

This is based on perf testing that indicates ~7000Mi should be able to
handle ~500k, and the max we currently see in prod over the last couple
of nights was ~5600Mi.

The DB pool size was reduced to 3 because 10 is excessive. 3 provides
one for message processing, one for ad-hoc operations via turnpike, and
1 spare.

Given stage is sized a bit smaller than prod, I'm leaving stage memory
alone, and leaving the lower default CPU request alone as well.

I expect this should result in stage:

* 1 less db connection
* 700m more CPU request (350m -> 1050m)
* 1500m more CPU limit (1500m -> 3)