SWATCH-32: Set swatch-tally-service to 3 replicas in prod
cpu request 2 -> 1
cpu limit 4.8 -> 1

This is based on the observation that the tally service doesn't
generally use more than 50% of given CPU.

memory request: 14000Mi -> 7000Mi
memory limit: 15488Mi -> 7000Mi

This is based on perf testing that indicates ~7000Mi should be able to
handle ~500k, and the max we currently see in prod over the last couple
of nights was ~5600Mi.

The DB pool size was reduced to 3 because 10 is excessive. 3 provides
one for message processing, one for ad-hoc operations via turnpike, and
1 spare.

I expect this should result in prod:

* 1 less db connection
* 1 more CPU request (2000m -> 3000m total)
* 1800m less CPU limit (4800m -> 3000m total)
* 7000Mi more memory request (14000Mi to 21000Mi total)
* 5512Mi more memory limit (15488Mi to 21000Mi total)