Disable the DVO checks and subsequent aleterting for MAS SSO
MAS SSO the core functionality has been migrated to
sso.redhat.com, we are in the process of migrating the remaining
functionality to other services such as auth.redhat.com. Currently
only the RHOAS-Kafka-Sre realm is active in MAS SSO, that is used for
data plane access (IDP) and AWS cloudwatch (For dataplane clusters).
We dont want the MAS SSO JIRAs created at this point for DVO checks.