image-builder: worker 5xx alerts
Image builder has added a new metric upstream to track
requests made to the worker. Previously, these metrics were
not being monitored. With this change, we are adding new
alerts to ensure that requests made to the worker server
are passing.

Alerts have been set to a lower severity and will be monitored
for the next 7d.