Monitor vault prod from internal stage cluster
Working on having monitoring for vault internal instances. In this case
we add monitor to vault prod from the only other internal cluster that
we have: appsres03ue1. This is far from ideal, but it is the only way we
have to have a monitor that is external. I'm leaving both probes (from
appsrep05ue1 and appsres03ue1) due to this special case. I don't fully
trust a probe from a stage cluster but it is better than only having a
probe from the same cluster.

part of APPSRE-5745

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>