Bump tally memory to 22000Mi and limit hbi systems to 250k
We've been dealing with an issue where a single customer causes OOMs
with growing numbers of system records. To mitigate the risk of OOMs, we
want to set a hard threshold, but also bump it just a bit higher than
the customer's current system count.

See https://issues.redhat.com/browse/SWATCH-782 for more details