feature: disable Floorist for Notifications
Notifications is integrated with Floorist because this allows running
queries to fetch data and then visualize it in Tableau. The problem is
that apparently it is hard to get the Tableau licenses, and therefore
the consequence is that we are sending data with Floorist that never
gets visualized afterwards.

This is taking resources from our deployments, and it is basically
performing NO-OPs, so we have decided to disable Floorist to save those
resources.