Add new splunk-audit-exporter deployment.
This deployment (compared to the splunk-forwarder-operator) will only be
deployed on hypershift management clusters.

To achieve this, it is deployed as a SelectorSyncSet to all hives, and
from there as a Policy to all hives, from there as a Policy to all
service-clusters.

The service-clusters in turn deploy the actual policy to management
clusters via a placement rule and the policy itself only applies to
hosted controlplane namespaces.