workaround fix KFM Grafana Dashboards SLOs title conflicts
Due to now Grafana KFM dashboards are referenced in Grafana
shared-resources this means that each dashboard is deployed in both
AppSRE Grafana prod and stage instances. This led to title collision
conflicts.
As a workaround we add the environment as a suffix of the dashboard
titles.
In the near future we will move the KFM dashboard definitions
of AppInterface and we will deploy the dashboards only on one Grafana
instance. In that scenario we will revert back the name to remove the
environment suffix.