remodel external-resources-maintainer change-type
the current external-resources-maintainer change-type was missing the restriction on `externalResources` that belonged to owned AWS accounts. this context is not available within this change-type because it does not know about its involvement in ownership expansion from `aws-account-owner`.

so the restrictive access to `externalResources` of a namespace need to be defined in the `aws-account-owner` change-type directly, because we have access to the account context and can use it to narrow the allowed change paths for external resources

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>