temporarily rename kfm metrics grafana dashboard configmap
The reason for this change is due to we are migrating KFM
dashboard definitions outside of the app-interface.
We want to keep the dashboard accessible during the migration process
and at the same time we want to avoid having clashes at id, uid, title
and ConfigMap level. Therefore to avoid the conflict we change the
ConfigMap on app-interface side