rename KFM grafana dashboard stage SLOs configmap name
The reason for this change is due to we are migrating KFM
dashboard definitions outside of the app-interface.
We want to keep the dashboard accessible during the migration process
and at the same time we want to avoid having clashes at id, uid, title
and ConfigMap level.
Therefore for the case of the stage SLOs we change the configmap name