[production] fix multi-images names
Now that we upgraded from v2.13.z to v2.15.z, we should be able to
rename those versions so that it will keep the same convention as in
``data/default_release_images.json`` in assisted-service repo.

The automatic bump job can then be unified to a single logic
(updating all entries that do not start with x.y.z).