image-builder: Switch the spot allocation strategy to capacity-optimized
Using lowest-price strategy for our production ASG causes AWS to pick up
instance types that are killed after 5 minutes. This breaks about 22% of
image builds according to our metrics. Let's switch to capacity-optimized
strategy that should allocate instances with lower chance of being
interrupted.

Long term, we would probably like to switch to price-capacity-optimized
which combines capacity-optimized with keep the price as low as possible,
but that's currently not possible in app-intf.

See:
https://docs.aws.amazon.com/autoscaling/ec2/userguide/ec2-auto-scaling-mixed-instances-groups.html#spot-allocation-strategy
https://github.com/app-sre/qontract-schemas/pull/400