Increment SWATCH db-changelog-cleanup job number to create new job
The Liquibase change management tool SWATCH uses for database migrations
maintains a table that holds a lock.  If the lock is held, other
instances of the migration will not run.  On occasion, the lock is not
released (the reason for this is unknown) and we need to run a job to
manually clear the lock.