saas-hive: set hash_length to 10
We hit a case where the first 7c of the commit sha was ambiguous, so our
appsre build script helpfully used 8c instead. Unfortunately, the
tooling on the appsre saas deploy side doesn't have the ability to do
this same disambiguation, so it tried the 7c prefix and the deploy
failed.

Increase the prefix length to 10c to greatly reduce (but not eliminate,
sad-face) the probability of future collisions.

HIVE-2194