Reduce idleMinutes to app-interface workers
The default is 30 minutes. The problem with this setting is that once we
have a new worker, Jenkins will try to round robin between them all so
it's easy that we have underused workers. Since we have the protection
of minSpareSize (the amount of idle workers at any given time), we can
be more aggressive with this setup so that we make sure we don't have
too many workers.

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>