Change stage default to full-iso
Prod defaults to full-iso so we should be the same in stage.
Also, this is causing issues when trying to create a cluster using the
UI with s390x architecture since it doesn't support full ISO.

The UI relies on the default value when creating the infraenv so the
default needs to be compatible with any architecture the user can
select.

This is a quick fix to unblock testing in stage. We should make the
default selection aware of the support matrix as a more resilient
solution.