Allow for release candidate installing for q-cli
Since we don't have a proper release process, the current way of
telling people to install things is broken since it relies in whatever
tag was created at some point. We publish to PyPi with every push to
master release candidate, people should use that instead of the latest
stable.

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>