remove checksum from the shan user GPG key
When exported, we require the GPG/PGP keys to only be Base64 encoded
rather than exported using an ASCII-armoured format. The latter adds a
CRC24 checksum at the end of the exported key, just after the Base64
payload. This is something we do not support currently and needs to be
removed.

Signed-off-by: Sébastien Han <seb@redhat.com>