OTA-930: Prepare Envoy config for integration OSUS
Doc: https://gitlab.cee.redhat.com/service/app-interface/-/blob/master/docs/app-sre/sop/add-route-for-ocm-component.md

OHSS-21550 should result in a domain for us to use instead of TODOs. The
Envoy config boilerplate already existed for OSUS but was identical to
stage. It seems the only thing that needs to change is the destination
endpoint.