Remove medium severity alert about Jenkins memory
JVM memory usage is a chainsaw due to GC, so it can be normal to be
under 25% for a few minutes. We could add a larger `for` condition but
at this point we have seen this alert so many times that we routinely
ignore it as it has become noise. this patch removes it as we have
another one that will page us if memory goes below 10%.

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>