Promote hive 7735bf34849d90f20217e9fc791595ca9e41a3ec
https://github.com/openshift/hive/compare/928bef3a88c299ac96a5c8fd4eba4f1e92ccd2c1...7735bf34849d90f20217e9fc791595ca9e41a3ec

---
```
commit: 0350613841566b4f892afc8de8eaf9e2a85294a2
author: Eric Fried
ClusterProvision.Spec.Metadata => .MetadataJSON

ClusterProvision.Spec.Metadata is of type runtime.RawExtension. It
doesn't work (maybe it never did) -- we think because a
runtime.RawExtension can only hold a (marshaled) runtime.Object, and the
installer's ClusterMetadata type isn't one.

This commit creates a new field, MetadataJSON, of type `[]byte`, where
we store the ClusterMetadata as a raw JSON string. All uses of the old
Metadata field are replaced with MetadataJSON.

Needed by HIVE-2204

commit: a1fa1b165d3c844e5665f0f34de995db416df226
author: Eric Fried
Azure custom resource group support

Day 0, the installer knows how to accept a custom resource group from
the install-config and use it for the cluster's cloud resources.
However, hive previously had no visibility to that field (in the
install-config or elsewhere); so for day 2 operations including
hibernation and additional MachinePools as well as for deprovisioning,
we were using the default (<cluster-infra-id>-rg). For machine pools,
this resulted in failures in the operation. In hibernation and
deprovisioning, the operation would "succeed" but no-op. For hibernation
this meant that machines were still running (incurring the associated
costs) even though the cluster claimed it was hibernating. For
deprovisioning, this meant the cloud resources were leaked.

This commit adds a new ClusterDeployment.Spec.ClusterMetadata section
for `Azure` containing a `ResourceGroupName` field, set as follows:
1. When provisioning, we discover it from the installer metadata during
installation.
2. When retrofitting a ClusterDeployment that was adopted or created
before custom Azure resource group support, we first attempt to discover
that resource group via the Infrastructure object in the remote cluster.
3. If that fails, we try parsing the install-config referenced by
ClusterDeployment.Spec.Provisioning.InstallConfigSecretRef.

Note that we will assign the default value if we're using the installer
metadata or the install-config and it's unset there. In any other case
where we can't figure it out (e.g. we can't find the infrastructure
object; or we can't parse the install-config) we just don't set it. The
consequence of that is...

This field is now required for any of the operations that use it -- if
it is absent, those operations will fail. While we don't expect this
condition to persist, as a fallback the user can manually set the value
if necessary.

HIVE-2204

commit: f1f63cbfad47b78d69f9a8f98aa481ae0a2053aa
author: Eric Fried
Only Watch Proxy when running on OpenShift

HIVE-2198 / #1999 / 253408ed introduced a `Watch()` on `Proxy` so we can
update environment variables and CA bundles when it changes. But `Proxy`
is an OpenShift CRD, so that `Watch()` blows up the operator when
running on vanilla k8s (and similar).

This commit conditions that `Watch()` such that we only add it if we're
running on OpenShift.

HIVE-2210

commit: bf502af9e5b4911b188f85213997b3bceebac616
author: Eric Fried
Move OpenShift discovery earlier, and cache it

We support running hive on non-OpenShift clusters, but we do a couple of
things differently if that's the case. Previously we were dynamically
discovering OpenShift-ness in the places we needed to know. With this
commit, we discover it at init time, and cache the result in the
reconcile object so we can consult it later.

Needed for HIVE-2210

```