OTA-933: Use standard dedicated-admin token for ota-stage
Previous `ota-stage` was using a manually configured, service-coupled
`cincinnati-integration/app-interface` ServiceAccount token with
manually crafted RBAC settings (that were deemed sufficient for
app-interface to manage only that service).

With better understanding of app-interface and further changes necessary
(like installing cert-manager on ota-stage) we need app-interface to
perform a wider set of operations, so we want to switch the
`automationToken` to use a more standard `app-sre-bot` SA created in the
`dedicated-admin` namespace [1] which should give it all necessary
permissions expected by any app-interface interaction via
`automationToken`. This also makes ota-stage less "special" from
app-interface point-of-view, which is likely a good thing.

[1] https://gitlab.cee.redhat.com/service/app-interface/-/blob/master/docs/app-sre/sop/app-interface-onboard-cluster.md#step-2-bot-access-and-app-sre-project-template